pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pompeii-agent"
dynamic = ["version"]
description = "Microkernel-style agent runtime — Pompeii-Agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
    "mcp>=1.2.0",
    "tiktoken>=0.5.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
pompeii-http = "app.http_runtime:main"
pompeii-cli = "app.cli_runtime:main"
pompeii-resource-migrate = "app.resource_migrate_runtime:main"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pompeii_agent = []
platform_layer = ["resources/**/*.yaml", "resources/**/*.db"]

[tool.setuptools.dynamic]
version = {attr = "app.version.__version__"}
